adding features

+++ src/main/java/com/code2ever/shoppinglist/services/category/CategoryService.java
@@ -5,7 +5,10 @@ import com.code2ever.shoppinglist.api.rest.model.CrudRestOperations;
 import com.code2ever.shoppinglist.api.rest.category.JsonCategory;
 import com.code2ever.shoppinglist.api.util.UtilClass;
 import com.code2ever.shoppinglist.model.category.Category;
+import com.code2ever.shoppinglist.model.item.Item;
+import com.code2ever.shoppinglist.model.list.ShoppingList;
 import com.code2ever.shoppinglist.repository.category.CategoryRepository;
+import com.code2ever.shoppinglist.repository.item.ItemRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -18,8 +21,11 @@ public class CategoryService implements CrudRestOperations<JsonCategory> {
 
     private final CategoryRepository repository;
 
-    public CategoryService(CategoryRepository repository) {
+    private final ItemRepository itemRepository;
+
+    public CategoryService(CategoryRepository repository, ItemRepository itemRepository) {
         this.repository = repository;
+        this.itemRepository = itemRepository;
     }
 
     @Override
@@ -60,6 +66,17 @@ public class CategoryService implements CrudRestOperations<JsonCategory> {
     @Override
     public void restDelete(Long id) {
         Objects.requireNonNull(id,"Id cant be null");
+        Category category = getCategoryById(id);
+        List<Item> itemList = itemRepository.getItemsByCategory(category);
+        if(!itemList.isEmpty()){
+            throw new ApplicationBusinessException("The category cant be deleted because there is items with this " +
+                    "category");
+        }
         repository.deleteById(id);
+    }
+
+    private Category getCategoryById(Long id) {
+        return repository.findById(id).orElseThrow(() -> new ApplicationBusinessException("There is no category with the id" +
+                id));
     }
 }
